Move processor implementation (#39)
* Move processor implementation

* Fix truncated comment

* Update generated clients

* Fix test compilation

* Tweak processor methods visibility

* Remove duplicated functions

* Change instruction name (#40)

* Change instruction name

* Update instruction name enum

* Update generated code

* Add helpers

* Update generated clients

* Update test import

* Update generated clients

---------
