+++ clients/js/test/_setup.ts
@@ -11,7 +11,7 @@ import {
 } from '@metaplex-foundation/umi';
 import { createUmi as baseCreateUmi } from '@metaplex-foundation/umi-bundle-tests';
 import {
-  DecompressableState,
+  DecompressibleState,
   MetadataArgsArgs,
   createTree as baseCreateTree,
   mintV1 as baseMintV1,
@@ -20,7 +20,7 @@ import {
   findTreeConfigPda,
   hashLeaf,
   mplBubblegum,
-  setDecompressableState,
+  setDecompressibleState,
 } from '../src';
 
 export const createUmi = async (endpoint?: string, airdropAmount?: SolAmount) =>
@@ -40,12 +40,12 @@ export const createTree = async (
     ...input,
   });
   builder = builder.append(
-    setDecompressableState(context, {
+    setDecompressibleState(context, {
       treeConfig:
         input.treeConfig ??
         findTreeConfigPda(context, { merkleTree: merkleTree.publicKey }),
       treeCreator: input.treeCreator ?? context.identity,
-      decompressableState: DecompressableState.Enabled,
+      decompressableState: DecompressibleState.Enabled,
     })
   );
   await builder.sendAndConfirm(context);
