+++ programs/bubblegum/program/src/processor/verify_creator.rs
@@ -1,0 +1,49 @@
+use anchor_lang::prelude::*;
+use spl_account_compression::{program::SplAccountCompression, Noop};
+
+use crate::{
+    processor::process_creator_verification,
+    state::{metaplex_adapter::MetadataArgs, TreeConfig},
+};
+
+#[derive(Accounts)]
+pub struct CreatorVerification<'info> {
+    #[account(
+        seeds = [merkle_tree.key().as_ref()],
+        bump,
+    )]
+    pub tree_authority: Account<'info, TreeConfig>,
+    /// CHECK: This account is checked in the instruction
+    pub leaf_owner: UncheckedAccount<'info>,
+    /// CHECK: This account is checked in the instruction
+    pub leaf_delegate: UncheckedAccount<'info>,
+    #[account(mut)]
+    /// CHECK: This account is modified in the downstream program
+    pub merkle_tree: UncheckedAccount<'info>,
+    pub payer: Signer<'info>,
+    pub creator: Signer<'info>,
+    pub log_wrapper: Program<'info, Noop>,
+    pub compression_program: Program<'info, SplAccountCompression>,
+    pub system_program: Program<'info, System>,
+}
+
+pub(crate) fn verify_creator<'info>(
+    ctx: Context<'_, '_, '_, 'info, CreatorVerification<'info>>,
+    root: [u8; 32],
+    data_hash: [u8; 32],
+    creator_hash: [u8; 32],
+    nonce: u64,
+    index: u32,
+    message: MetadataArgs,
+) -> Result<()> {
+    process_creator_verification(
+        ctx,
+        root,
+        data_hash,
+        creator_hash,
+        nonce,
+        index,
+        message,
+        true,
+    )
+}
